--- alpine/views.py
@@ -1,20 +1,0 @@
-from django.urls import reverse_lazy
-from django.shortcuts import render
-from django.views.generic import CreateView
-from django.contrib.auth.forms import UserCreationForm
-
-
-
-# Create your views here.
-
-def index(req):
-    ctx = {}
-    ctx["title"] = "Alpine Index"
-    ctx["content"] = "alpine index"
-
-    return render(req, "pages/example.html", ctx)
-
-class SignUpView(CreateView):
-    form_class = UserCreationForm
-    success_url = reverse_lazy("login")
-    template_name = "registration/signup.html"
