+++ api/views.py
@@ -1,4 +1,9 @@
+from django.urls import reverse_lazy
 from django.shortcuts import render
+from django.views.generic import CreateView
+from django.contrib.auth.forms import UserCreationForm
+
+
 
 # Create your views here.
 
@@ -7,3 +12,8 @@ def index(req):
     ctx["title"] = "Alpine API Index"
     ctx["content"] = "api index"
     return render(req, "pages/example.html", ctx)
+
+class SignUpView(CreateView):
+    form_class = UserCreationForm
+    success_url = reverse_lazy("login")
+    template_name = "registration/signup.html"
